+++ Conquer/src/org/jel/game/data/strategy/StrategyProvider.java
@@ -3,7 +3,7 @@ package org.jel.game.data.strategy;
 public interface StrategyProvider {
 	Strategy buildStrategy();
 
-	byte getId();
+	int getId();
 
 	String getName();
 }
